Deduplicate common instructions, reduce bloat
