+++ mlir/include/mlir/Dialect/XeGPU/uArch/uArchBase.h
@@ -7,19 +7,28 @@
 //===----------------------------------------------------------------------===//
 //
 // \file
-// Base uArch definition for different architectures.
-//
+// Base uArch definition for different architectures, plus the SPIRV / Khronos
+// OpenCL extension instruction defaults shared across Intel Xe uArchs.
 //
 //===----------------------------------------------------------------------===//
 #ifndef MLIR_DIALECT_XEGPU_UARCH_UARCHBASE_H
 #define MLIR_DIALECT_XEGPU_UARCH_UARCHBASE_H
 
-#include <shared_mutex>
+#include <cassert>
+#include <optional>
 #include <tuple>
+#include <utility>
 
+#include "mlir/IR/BuiltinTypes.h"
+#include "mlir/IR/TypeUtilities.h"
 #include "mlir/IR/Types.h"
+#include "llvm/ADT/DenseMap.h"
+#include "llvm/ADT/STLExtras.h"
 #include "llvm/ADT/SmallVector.h"
+#include "llvm/ADT/StringRef.h"
 #include "llvm/Support/Casting.h"
+#include "llvm/Support/DebugLog.h"
+#include "llvm/Support/ErrorHandling.h"
 
 namespace mlir {
 namespace xegpu {
@@ -38,7 +47,6 @@ enum class InstructionKind {
   Subgroup2DBlockPrefetch, // Subgroup-level 2D block prefetch instruction
   StoreScatter,            // Lane-level store (scalar, vector)
   LoadGather,              // Lane-level load (scalar, vector)
-  // @TODO: Add more instructions as needed
 };
 
 // A struct to represent basic information about an instruction.
@@ -80,63 +88,6 @@ protected:
                                   // subgroup, workgroup, cluster)
 };
 
-enum class RegisterFileMode : uint8_t { Small, Large };
-enum class RegisterFileType : uint8_t { GRF, ARF };
-
-// A struct to represent register file information
-struct RegisterFileInfo {
-  // Constructor
-  RegisterFileInfo() = default;
-  RegisterFileInfo(uint32_t size,
-                   const llvm::SmallVector<RegisterFileMode, 4> &mode,
-                   const llvm::SmallVector<uint32_t, 4> &numRegs)
-      : size(size), mode(mode), numRegsPerThreadPerMode(numRegs) {}
-
-  // Get methods
-  uint32_t getSize() const { return size; }
-
-  const llvm::SmallVector<RegisterFileMode, 4> &getModes() const {
-    return mode;
-  }
-
-  const llvm::SmallVector<uint32_t, 4> &getNumRegsPerThreadPerMode() const {
-    return numRegsPerThreadPerMode;
-  }
-
-protected:
-  uint32_t size; // size per register in bits
-  llvm::SmallVector<RegisterFileMode, 4>
-      mode; // e.g., "small", "large" GRF modes
-  llvm::SmallVector<uint32_t, 4>
-      numRegsPerThreadPerMode; // number of registers per thread per mode
-};
-
-enum class CacheHierarchyLevel { L1 = 1, L2 = 2, L3 = 3 };
-
-// A struct to represent cache information
-struct CacheInfo {
-  // Constructor
-  CacheInfo() = default;
-  CacheInfo(uint32_t size, uint32_t line_size,
-            CacheHierarchyLevel hierarchy_level)
-      : size(size), line_size(line_size), hierarchy_level(hierarchy_level) {}
-
-  virtual ~CacheInfo() = default;
-
-  // Get methods
-  uint32_t getSize() const { return size; }
-  uint32_t getLineSize() const { return line_size; }
-  CacheHierarchyLevel getHierarchyLevel() const { return hierarchy_level; }
-
-protected:
-  uint32_t size;
-  uint32_t line_size;
-  CacheHierarchyLevel hierarchy_level;
-  // @TODO: Add more fields as needed (e.g., associativity, num_banks,
-  // bank_size, num_ports, port_width, bank_conflicts, hierarchy_level,
-  // latency, throughput, bandwidth)
-};
-
 struct uArch {
   enum class Kind {
     // Xe2 family
@@ -157,33 +108,7 @@ struct uArch {
   }
   virtual ~uArch() = default;
   Kind getKind() const { return kind; }
-  StringRef getName() const { return getUArchName(kind); }
-  StringRef getDescription() const { return getUArchDescription(kind); }
 
-  static StringRef getUArchName(Kind k) {
-    switch (k) {
-    case Kind::PVC:
-      return "pvc";
-    case Kind::BMG:
-      return "bmg";
-    case Kind::CRI:
-      return "cri";
-    default:
-      return "";
-    }
-  }
-
-  static StringRef getUArchDescription(Kind k) {
-    switch (k) {
-    case Kind::PVC:
-      return "Ponte Vecchio Architecture";
-    case Kind::BMG:
-      return "Battlemage Architecture";
-    case Kind::CRI:
-      return "Crescent Island Architecture";
-    }
-    llvm_unreachable("Unknown uArch::Kind");
-  }
   virtual int getSubgroupSize() const = 0;
   virtual unsigned getGeneralPackedFormatBitSize() const = 0;
 
@@ -202,35 +127,6 @@ protected:
   Kind kind;
   llvm::SmallDenseMap<InstructionKind, const Instruction *, 32>
       instructionRegistry;
-};
-
-// A struct to represent shared memory information
-struct SharedMemory {
-  // Constructor
-  SharedMemory(uint32_t size, uint32_t alignment)
-      : size(size), alignment(alignment) {}
-
-  // Get methods
-  uint32_t getSize() const { return size; }
-  uint32_t getAlignment() const { return alignment; }
-
-protected:
-  uint32_t size;      // in bytes
-  uint32_t alignment; // in bytes
-  // @TODO: Add more fields as needed (e.g., latency, throughput, bandwidth)
-};
-
-struct XeCoreInfo {
-  uint32_t num_threads;
-  SharedMemory shared_memory;
-  uint32_t num_vector_units;
-  uint32_t num_matrix_units;
-
-  XeCoreInfo(uint32_t num_threads, const SharedMemory &shared_memory,
-             uint32_t num_vector_units, uint32_t num_matrix_units)
-      : num_threads(num_threads), shared_memory(shared_memory),
-        num_vector_units(num_vector_units), num_matrix_units(num_matrix_units) {
-  }
 };
 
 //===----------------------------------------------------------------------===//
@@ -256,19 +152,7 @@ struct MMAInstructionInterface {
   // this method.
   virtual llvm::SmallVector<Type, 8>
   getSupportedTypes(MLIRContext &context, MMAOpndKind matrixType) = 0;
-  virtual bool
-  checkSupportedShapesAndTypes(std::pair<uint32_t, uint32_t> AShape,
-                               std::pair<uint32_t, uint32_t> BShape,
-                               std::pair<uint32_t, uint32_t> CShape,
-                               std::pair<uint32_t, uint32_t> DShape, Type AType,
-                               Type BType, Type CType, Type DType) = 0;
-  virtual bool checkSupportedTypes(Type AType, Type BType, Type CType,
-                                   Type DType) = 0;
-  virtual bool validate(std::pair<uint32_t, uint32_t> AShape,
-                        std::pair<uint32_t, uint32_t> BShape,
-                        std::pair<uint32_t, uint32_t> CShape,
-                        std::pair<uint32_t, uint32_t> DShape, Type AType,
-                        Type BType, Type CType, Type DType) = 0;
+
   virtual llvm::SmallVector<uint32_t, 8> getSupportedM(Type type) const = 0;
   virtual llvm::SmallVector<uint32_t, 8> getSupportedK(Type type) const = 0;
   virtual llvm::SmallVector<uint32_t, 8> getSupportedN(Type type) const = 0;
@@ -282,51 +166,425 @@ struct MMAInstructionInterface {
 // transform / transpose / upConv flags are only meaningful for loads; store
 // and prefetch implementations ignore them.
 struct BlockIOInstructionInterface {
+  using BlockShapes =
+      std::tuple<llvm::ArrayRef<int>, llvm::ArrayRef<int>, llvm::ArrayRef<int>>;
+
   // Returns the supported (widths, heights, counts) for the given element
   // type, or std::nullopt if the element type is unsupported.
-  virtual std::optional<
-      std::tuple<llvm::ArrayRef<int>, llvm::ArrayRef<int>, llvm::ArrayRef<int>>>
+  std::optional<BlockShapes>
   getBlockWidthHeightCount(Type elemTy, bool hasTransform = false,
                            bool hasTranspose = false,
-                           bool upConv = false) const = 0;
+                           bool upConv = false) const {
+    return computeBlockWidthHeightCount(elemTy, hasTransform, hasTranspose,
+                                        upConv);
+  }
+
   // Bit size of the packed format used by this block instruction.
   virtual int32_t getPackedFormatBitSize() const = 0;
   virtual ~BlockIOInstructionInterface() = default;
+
+protected:
+  virtual std::optional<BlockShapes>
+  computeBlockWidthHeightCount(Type elemTy, bool hasTransform,
+                               bool hasTranspose, bool upConv) const = 0;
 };
 
 //===----------------------------------------------------------------------===//
-// Common instructions (shared across architectures)
+// Common virtual ISA instructions (shared across architectures)
 //===----------------------------------------------------------------------===//
 
-struct LoadGatherInstructionInterface : public Instruction {
-  LoadGatherInstructionInterface()
-      : Instruction(InstructionKind::LoadGather, InstructionScope::Lane) {}
+//===----------------------------------------------------------------------===//
+// SPIRV
+//===----------------------------------------------------------------------===//
+template <InstructionKind Kind>
+struct ScatterIoInstructionInterface : public Instruction {
+  static_assert(Kind == InstructionKind::LoadGather ||
+                    Kind == InstructionKind::StoreScatter,
+                "ScatterIO only supports LoadGather / StoreScatter");
+
+  ScatterIoInstructionInterface() : Instruction(Kind, InstructionScope::Lane) {}
+
   static bool classof(const Instruction *B) {
-    return B->getInstructionKind() == InstructionKind::LoadGather;
+    return B->getInstructionKind() == Kind;
   }
 
-  virtual int32_t getMaxLaneLoadSize(int32_t bitWidth) const = 0;
-  virtual ~LoadGatherInstructionInterface() = default;
+  virtual int32_t getMaxLaneAccessSizeBytes() const = 0;
+  virtual ~ScatterIoInstructionInterface() = default;
+};
+struct LoadGatherInstruction
+    : public ScatterIoInstructionInterface<InstructionKind::LoadGather> {
+  int32_t getMaxLaneAccessSizeBytes() const override { return 16; }
 };
 
-struct StoreScatterInstructionInterface : public Instruction {
-  StoreScatterInstructionInterface()
-      : Instruction(InstructionKind::StoreScatter, InstructionScope::Lane) {}
+struct StoreScatterInstruction
+    : public ScatterIoInstructionInterface<InstructionKind::StoreScatter> {
+  int32_t getMaxLaneAccessSizeBytes() const override { return 16; }
+};
+
+//===----------------------------------------------------------------------===//
+// SPIRV / OpenCL-extension subgroup instructions
+//
+// These come from cl_intel_subgroup_2d_block_io and
+// cl_intel_subgroup_matrix_multiply_accumulate. A uArch only needs to
+// subclass when it diverges from the extension defaults.
+//===----------------------------------------------------------------------===//
+
+struct Subgroup2DBlockStoreInstruction : public Instruction,
+                                         public BlockIOInstructionInterface {
+  Subgroup2DBlockStoreInstruction()
+      : Instruction(InstructionKind::Subgroup2DBlockStore,
+                    InstructionScope::Subgroup) {}
   static bool classof(const Instruction *B) {
-    return B->getInstructionKind() == InstructionKind::StoreScatter;
+    return B->getInstructionKind() == InstructionKind::Subgroup2DBlockStore;
+  }
+  // Source :
+  // https://registry.khronos.org/OpenCL/extensions/intel/cl_intel_subgroup_2d_block_io.html#_add_a_new_section_5_2_x_cl_intel_subgroup_2d_block_io
+  // Stores ignore the transform / transpose / upConv flags.
+  int32_t getPackedFormatBitSize() const override { return 16; }
+
+protected:
+  std::optional<BlockShapes>
+  computeBlockWidthHeightCount(Type elemTy, bool /*hasTransform*/,
+                               bool /*hasTranspose*/,
+                               bool /*upConv*/) const override {
+    static const int kHeight[] = {1, 2, 4, 8};
+    static const int kWidth16[] = {16};
+    static const int kCount[] = {1};
+    const int elemByteSize = elemTy.getIntOrFloatBitWidth() / 8;
+    if (elemByteSize == 1 || elemByteSize == 2 || elemByteSize == 4)
+      return std::make_tuple(llvm::ArrayRef<int>(kWidth16),
+                             llvm::ArrayRef<int>(kHeight),
+                             llvm::ArrayRef<int>(kCount));
+    return std::nullopt;
+  }
+};
+
+struct Subgroup2DBlockLoadInstruction : public Instruction,
+                                        public BlockIOInstructionInterface {
+  Subgroup2DBlockLoadInstruction()
+      : Instruction(InstructionKind::Subgroup2DBlockLoad,
+                    InstructionScope::Subgroup) {}
+  static bool classof(const Instruction *B) {
+    return B->getInstructionKind() == InstructionKind::Subgroup2DBlockLoad;
   }
 
-  virtual int32_t getMaxLaneStoreSize(int32_t bitWidth) const = 0;
-  virtual ~StoreScatterInstructionInterface() = default;
+  // Source :
+  // https://registry.khronos.org/OpenCL/extensions/intel/cl_intel_subgroup_2d_block_io.html#_add_a_new_section_5_2_x_cl_intel_subgroup_2d_block_io
+  int32_t getPackedFormatBitSize() const override { return 16; }
+
+protected:
+  std::optional<BlockShapes>
+  computeBlockWidthHeightCount(Type elemTy, bool hasTransform,
+                               bool hasTranspose, bool upConv) const override {
+    static const int kHeightAtLeast1[] = {1, 2, 4, 8, 16, 32};
+    static const int kHeightAtLeast8[] = {8, 16, 32};
+    static const int kHeightAtLeast16[] = {16, 32};
+    static const int kHeight32[] = {32};
+    static const int kHeight64[] = {64};
+
+    static const int kWidth64[] = {64};
+    static const int kWidth32[] = {32};
+    static const int kWidth16[] = {16};
+    static const int kWidthAtLeast16[] = {16, 32};
+    static const int kWidthAtLeast32[] = {32, 64};
+    static const int kWidth8[] = {8};
+
+    static const int32_t kCount1[] = {1};
+    static const int32_t kCount2[] = {1, 2};
+    static const int32_t kCount4[] = {1, 2, 4};
+    static const int32_t kCount4Only[] = {4};
+    // (elemBits, transform, transpose, upConvert)
+    using Key = std::tuple<int, uint8_t, uint8_t, uint8_t>;
+    // (widths, heights, counts)
+    using Value = std::tuple<llvm::ArrayRef<int32_t>, llvm::ArrayRef<int32_t>,
+                             llvm::ArrayRef<int32_t>>;
+    // The table is keyed on element bit width so sub-byte elements can be
+    // expressed directly. 4-bit elements are packed two-per-byte, so their
+    // widths (or heights, when transformed) are double the 8-bit rows.
+    static const llvm::DenseMap<Key, Value> kMap = {
+        {{8, false, false, false}, {kWidthAtLeast16, kHeightAtLeast1, kCount2}},
+        {{8, false, false, true}, {kWidth16, kHeightAtLeast8, kCount4Only}},
+        {{16, false, false, false}, {kWidth16, kHeightAtLeast1, kCount2}},
+        {{32, false, false, false}, {kWidth16, kHeightAtLeast1, kCount1}},
+        // Block Loads with Transform:
+        {{8, true, false, false}, {kWidth16, kHeight32, kCount4}},
+        {{16, true, false, false}, {kWidth16, kHeightAtLeast16, kCount2}},
+        // Block Loads with Transpose:
+        {{8, false, true, false}, {kWidth32, kHeightAtLeast16, kCount1}},
+        {{16, false, true, false}, {kWidth16, kHeightAtLeast16, kCount1}},
+        {{32, false, true, false}, {kWidth8, kHeightAtLeast16, kCount1}},
+        // 4-bit elements (sub-byte):
+        {{4, false, false, false}, {kWidthAtLeast32, kHeightAtLeast1, kCount2}},
+        {{4, false, false, true}, {kWidth32, kHeightAtLeast8, kCount4Only}},
+        {{4, true, false, false}, {kWidth16, kHeight64, kCount4}},
+        {{4, false, true, false}, {kWidth64, kHeightAtLeast16, kCount1}}};
+    int elemBitSize = elemTy.getIntOrFloatBitWidth();
+    auto it = kMap.find({elemBitSize, hasTransform, hasTranspose, upConv});
+    if (it != kMap.end())
+      return it->second;
+    return std::nullopt;
+  }
 };
 
-struct SpirvLoadGatherInstruction : public LoadGatherInstructionInterface {
-  int32_t getMaxLaneLoadSize(int32_t bitWidth) const override { return 16; }
+struct Subgroup2DBlockPrefetchInstruction : public Instruction,
+                                            public BlockIOInstructionInterface {
+  Subgroup2DBlockPrefetchInstruction()
+      : Instruction(InstructionKind::Subgroup2DBlockPrefetch,
+                    InstructionScope::Subgroup) {}
+  static bool classof(const Instruction *B) {
+    return B->getInstructionKind() == InstructionKind::Subgroup2DBlockPrefetch;
+  }
+  // Source :
+  // https://registry.khronos.org/OpenCL/extensions/intel/cl_intel_subgroup_buffer_prefetch.html#_add_a_new_section_6_15_x_sub_group_prefetch_functions
+  // Prefetches ignore the transform / transpose / upConv flags.
+  int32_t getPackedFormatBitSize() const override { return 16; }
+
+protected:
+  std::optional<BlockShapes>
+  computeBlockWidthHeightCount(Type elemTy, bool /*hasTransform*/,
+                               bool /*hasTranspose*/,
+                               bool /*upConv*/) const override {
+    static const int kHeightAtLeast1[] = {1, 2, 4, 8, 16, 32};
+
+    static const int kWidth32[] = {32};
+    static const int kWidth16[] = {16};
+
+    static const int32_t kCount1[] = {1};
+    static const int32_t kCount2[] = {1, 2};
+    // elemBytes
+    using Key = int;
+    // (widths, heights, counts)
+    using Value = std::tuple<llvm::ArrayRef<int32_t>, llvm::ArrayRef<int32_t>,
+                             llvm::ArrayRef<int32_t>>;
+    static const llvm::DenseMap<Key, Value> kMap = {
+        {1, {kWidth32, kHeightAtLeast1, kCount2}},
+        {2, {kWidth16, kHeightAtLeast1, kCount2}},
+        {4, {kWidth16, kHeightAtLeast1, kCount1}},
+    };
+    const int elemByteSize = elemTy.getIntOrFloatBitWidth() / 8;
+    auto it = kMap.find(elemByteSize);
+    if (it != kMap.end())
+      return it->second;
+    return std::nullopt;
+  }
 };
 
-struct SpirvStoreScatterInstruction : public StoreScatterInstructionInterface {
-  int32_t getMaxLaneStoreSize(int32_t bitWidth) const override { return 16; }
+struct SubgroupMatrixMultiplyAcc : public Instruction,
+                                   public MMAInstructionInterface {
+  SubgroupMatrixMultiplyAcc(unsigned packedFormatBitSizeA,
+                            unsigned packedFormatBitSizeB)
+      : Instruction(InstructionKind::SubgroupMatrixMultiplyAcc,
+                    InstructionScope::Subgroup),
+        packedFormatBitSizeA(packedFormatBitSizeA),
+        packedFormatBitSizeB(packedFormatBitSizeB) {}
+  static bool classof(const Instruction *B) {
+    return B->getInstructionKind() ==
+           InstructionKind::SubgroupMatrixMultiplyAcc;
+  }
+  // Source:
+  // https://registry.khronos.org/OpenCL/extensions/intel/cl_intel_subgroup_matrix_multiply_accumulate.html
+
+  llvm::SmallVector<std::pair<uint32_t, uint32_t>, 16>
+  getSupportedShapes(Type dataType, MMAOpndKind matrixType) override;
+  llvm::SmallVector<Type, 8> getSupportedTypes(MLIRContext &context,
+                                               MMAOpndKind matrixType) override;
+
+  llvm::SmallVector<uint32_t, 8> getSupportedM(Type type) const override;
+  llvm::SmallVector<uint32_t, 8> getSupportedK(Type type) const override;
+  llvm::SmallVector<uint32_t, 8> getSupportedN(Type type) const override;
+
+  unsigned getPackedFormatBitSizeA() const { return packedFormatBitSizeA; }
+  unsigned getPackedFormatBitSizeB() const { return packedFormatBitSizeB; }
+  bool isLaneLayoutRowMajorOrder() const override { return true; }
+
+protected:
+  const unsigned packedFormatBitSizeA;
+  const unsigned packedFormatBitSizeB;
 };
+
+struct SubgroupScaledMatrixMultiplyAcc : public Instruction,
+                                         public MMAInstructionInterface {
+  SubgroupScaledMatrixMultiplyAcc(unsigned packedFormatBitSizeA,
+                                  unsigned packedFormatBitSizeB)
+      : Instruction(InstructionKind::SubgroupScaledMatrixMultiplyAcc,
+                    InstructionScope::Subgroup),
+        packedFormatBitSizeA(packedFormatBitSizeA),
+        packedFormatBitSizeB(packedFormatBitSizeB) {}
+  static bool classof(const Instruction *B) {
+    return B->getInstructionKind() ==
+           InstructionKind::SubgroupScaledMatrixMultiplyAcc;
+  }
+  // Source:
+  // https://github.com/intel/llvm/blob/sycl/sycl/doc/design/spirv-extensions/SPV_INTEL_subgroup_scaled_matrix_multiply_accumulate.asciidoc
+
+  llvm::SmallVector<std::pair<uint32_t, uint32_t>, 16>
+  getSupportedShapes(Type dataType, MMAOpndKind matrixType) override;
+  llvm::SmallVector<Type, 8> getSupportedTypes(MLIRContext &context,
+                                               MMAOpndKind matrixType) override;
+
+  llvm::SmallVector<uint32_t, 8> getSupportedM(Type type) const override;
+  llvm::SmallVector<uint32_t, 8> getSupportedK(Type type) const override;
+  llvm::SmallVector<uint32_t, 8> getSupportedN(Type type) const override;
+
+  unsigned getPackedFormatBitSizeA() const { return packedFormatBitSizeA; }
+  unsigned getPackedFormatBitSizeB() const { return packedFormatBitSizeB; }
+  bool isLaneLayoutRowMajorOrder() const override { return true; }
+
+protected:
+  const unsigned packedFormatBitSizeA;
+  const unsigned packedFormatBitSizeB;
+};
+
+//===----------------------------------------------------------------------===//
+// Inline implementations
+//===----------------------------------------------------------------------===//
+
+namespace util {
+inline llvm::SmallVector<std::pair<uint32_t, uint32_t>, 16>
+crossProduct(const llvm::SmallVector<uint32_t, 8> &a,
+             const llvm::SmallVector<uint32_t, 8> &b) {
+  llvm::SmallVector<std::pair<uint32_t, uint32_t>, 16> result;
+  for (unsigned x : a)
+    for (unsigned y : b)
+      result.emplace_back(x, y);
+  return result;
+}
+} // namespace util
+
+inline llvm::SmallVector<std::pair<uint32_t, uint32_t>, 16>
+SubgroupMatrixMultiplyAcc::getSupportedShapes(Type dataType,
+                                              MMAOpndKind matrixType) {
+  auto M = getSupportedM(dataType);
+  auto K = getSupportedK(dataType);
+  auto N = getSupportedN(dataType);
+  switch (matrixType) {
+  case MMAOpndKind::MatrixA:
+    return util::crossProduct(M, K);
+  case MMAOpndKind::MatrixB:
+    return util::crossProduct(K, N);
+  case MMAOpndKind::MatrixC:
+  case MMAOpndKind::MatrixD:
+    return util::crossProduct(M, N);
+  }
+  return {};
+}
+
+inline llvm::SmallVector<Type, 8>
+SubgroupMatrixMultiplyAcc::getSupportedTypes(MLIRContext &context,
+                                             MMAOpndKind matrixType) {
+  Type bf16Type = BFloat16Type::get(&context);
+  Type f16Type = Float16Type::get(&context);
+  Type tf32Type = FloatTF32Type::get(&context);
+  Type f32Type = Float32Type::get(&context);
+
+  switch (matrixType) {
+  case MMAOpndKind::MatrixA:
+  case MMAOpndKind::MatrixB:
+    return {bf16Type, f16Type, tf32Type};
+  case MMAOpndKind::MatrixC:
+  case MMAOpndKind::MatrixD:
+    return {bf16Type, f16Type, f32Type};
+  }
+  return {};
+}
+
+inline llvm::SmallVector<uint32_t, 8>
+SubgroupMatrixMultiplyAcc::getSupportedM(Type type) const {
+  return {1, 2, 3, 4, 5, 6, 7, 8};
+}
+
+inline llvm::SmallVector<uint32_t, 8>
+SubgroupMatrixMultiplyAcc::getSupportedK(Type type) const {
+  assert(type.isIntOrFloat() && "Matrix type must be int or float");
+  auto bitWidth = type.getIntOrFloatBitWidth();
+  uint32_t kSize = 0;
+  switch (bitWidth) {
+  case 4:
+    kSize = 64;
+    break;
+  case 8:
+    kSize = 32;
+    break;
+  case 16:
+    kSize = 16;
+    break;
+  case 32:
+    kSize = 8;
+    break;
+  default:
+    llvm_unreachable("Invalid int or float");
+  }
+  return {kSize};
+}
+
+inline llvm::SmallVector<uint32_t, 8>
+SubgroupMatrixMultiplyAcc::getSupportedN(Type type) const {
+  return {16};
+}
+
+inline llvm::SmallVector<std::pair<uint32_t, uint32_t>, 16>
+SubgroupScaledMatrixMultiplyAcc::getSupportedShapes(Type dataType,
+                                                    MMAOpndKind matrixType) {
+  // Avoid calling getSupportedK for C/D types (which are f32/bf16
+  // and not valid for the K-dimension bit-width calculation).
+  switch (matrixType) {
+  case MMAOpndKind::MatrixA:
+    return util::crossProduct(getSupportedM(dataType), getSupportedK(dataType));
+  case MMAOpndKind::MatrixB:
+    return util::crossProduct(getSupportedK(dataType), getSupportedN(dataType));
+  case MMAOpndKind::MatrixC:
+  case MMAOpndKind::MatrixD:
+    return util::crossProduct(getSupportedM(dataType), getSupportedN(dataType));
+  }
+  return {};
+}
+
+inline llvm::SmallVector<Type, 8>
+SubgroupScaledMatrixMultiplyAcc::getSupportedTypes(MLIRContext &context,
+                                                   MMAOpndKind matrixType) {
+  Type f8E4M3FNType = Float8E4M3FNType::get(&context);
+  Type f8E5M2Type = Float8E5M2Type::get(&context);
+  Type f4E2M1FNType = Float4E2M1FNType::get(&context);
+  Type bf16Type = BFloat16Type::get(&context);
+  Type f32Type = Float32Type::get(&context);
+
+  switch (matrixType) {
+  case MMAOpndKind::MatrixA:
+  case MMAOpndKind::MatrixB:
+    return {f8E4M3FNType, f8E5M2Type, f4E2M1FNType};
+  case MMAOpndKind::MatrixC:
+  case MMAOpndKind::MatrixD:
+    return {bf16Type, f32Type};
+  }
+  return {};
+}
+
+inline llvm::SmallVector<uint32_t, 8>
+SubgroupScaledMatrixMultiplyAcc::getSupportedM(Type type) const {
+  return {8};
+}
+
+inline llvm::SmallVector<uint32_t, 8>
+SubgroupScaledMatrixMultiplyAcc::getSupportedK(Type type) const {
+  assert(type.isIntOrFloat() && "Matrix type must be int or float");
+  auto bitWidth = type.getIntOrFloatBitWidth();
+  switch (bitWidth) {
+  case 4:
+    return {64}; // FP4: scale K by 4 (base 16-bit K=16 -> 64)
+  case 8:
+    return {32}; // FP8: scale K by 2 (base 16-bit K=16 -> 32)
+  default:
+    // Scaled dpas only supports FP8 (8-bit) and FP4 (4-bit) types for A/B
+    // matrices. Return empty so callers can gracefully reject unsupported
+    // types instead of aborting.
+    return {};
+  }
+}
+
+inline llvm::SmallVector<uint32_t, 8>
+SubgroupScaledMatrixMultiplyAcc::getSupportedN(Type type) const {
+  return {16};
+}
 
 } // namespace uArch
 } // namespace xegpu
