+++ mlir/lib/Dialect/XeGPU/Transforms/XeGPUPeepHoleOptimizer.cpp
@@ -106,8 +106,9 @@ static bool canBeOptimizedForTranspose(xegpu::TensorDescType tdescType) {
 
 /// Check if a tensor desc type can be optimized for transpose, if so return the
 /// new optimized tensor desc type with a valid transpose layout.
-static xegpu::TensorDescType tryOptimize(xegpu::TensorDescType tdescType,
-                                         const uArch *targetuArch) {
+static xegpu::TensorDescType
+tryOptimize(xegpu::TensorDescType tdescType,
+            const xegpu::uArch::uArch *targetuArch) {
   if (!canBeOptimizedForTranspose(tdescType))
     return tdescType;
   auto laneData = getMaybeLaneData(tdescType)
@@ -124,8 +125,10 @@ static xegpu::TensorDescType tryOptimize(xegpu::TensorDescType tdescType,
   Type newElemTy = IntegerType::get(tdescType.getContext(), newBitWidth);
   // Supported shape is the max transpose shape that can be supported by
   // hardware that is less than or equal to required shape.
-  auto *blockLoadTarget = dyn_cast<Subgroup2DBlockLoadInstruction>(
-      targetuArch->getInstruction(InstructionKind::Subgroup2DBlockLoad));
+  auto *blockLoadTarget =
+      dyn_cast<xegpu::uArch::Subgroup2DBlockLoadInstruction>(
+          targetuArch->getInstruction(
+              xegpu::uArch::InstructionKind::Subgroup2DBlockLoad));
   auto maybeHWParams = blockLoadTarget->getBlockWidthHeightCount(
       newElemTy, /** has transform */ false, /** has transpose */ true);
   // If no HW params found, return the original type.
@@ -261,7 +264,7 @@ public:
             chipStr.value() == "cri") &&
            "Expecting target chip to be pvc, bmg or cri for transpose "
            "optimization.");
-    const uArch *targetuArch = xegpu::uArch::getUArch(chipStr.value());
+    const auto *targetuArch = xegpu::uArch::getUArch(chipStr.value());
 
     auto convertType = tryOptimize(tdescTy, targetuArch);
     if (convertType == tdescTy)
